# backend/open-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Coin bot Api",
    "description": "Describes the backend api",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Organisation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "ownerId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "updateDate": {
            "format": "date-time",
            "type": "string"
          },
          "deletedDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "ownerId",
          "name",
          "createdDate",
          "updateDate",
          "deletedDate"
        ]
      },
      "Person": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "blocked": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "familyName": {
            "type": "string"
          },
          "givenName": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "auth0UserId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "updateDate": {
            "format": "date-time",
            "type": "string"
          },
          "deletedDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "uuid",
          "emailVerified",
          "blocked",
          "picture",
          "auth0UserId",
          "createdDate",
          "updateDate",
          "deletedDate"
        ]
      },
      "UpdateOrganisationDto": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Person"
            }
          },
          "owner": {
            "$ref": "#/components/schemas/Person"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UpdatePersonDto": {
        "type": "object",
        "properties": {}
      },
      "TwitterUserMentionMeta": {
        "type": "object",
        "properties": {}
      },
      "TriggerTypeEnum": {
        "type": "string",
        "enum": [
          "TWITTER_USER_MENTION",
          "NO_ACTION_DEFAULT"
        ]
      },
      "Trigger": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "updateSchedule": {
            "type": "string"
          },
          "triggerType": {
            "$ref": "#/components/schemas/TriggerTypeEnum"
          },
          "triggerResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriggerResult"
            }
          },
          "customBotId": {
            "type": "number"
          },
          "meta": {
            "type": "object"
          },
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "updateDate": {
            "format": "date-time",
            "type": "string"
          },
          "deletedDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "updateSchedule",
          "triggerType",
          "triggerResults",
          "customBotId",
          "meta",
          "createdDate",
          "updateDate",
          "deletedDate"
        ]
      },
      "TriggerResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "result": {
            "type": "boolean"
          },
          "trigger": {
            "$ref": "#/components/schemas/Trigger"
          },
          "triggerId": {
            "type": "number"
          },
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "updateDate": {
            "format": "date-time",
            "type": "string"
          },
          "deletedDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "result",
          "trigger",
          "triggerId",
          "createdDate",
          "updateDate",
          "deletedDate"
        ]
      },
      "CreateTriggerDto": {
        "type": "object",
        "properties": {
          "updateSchedule": {
            "type": "string"
          },
          "triggerType": {
            "$ref": "#/components/schemas/TriggerTypeEnum"
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "updateSchedule",
          "triggerType",
          "meta"
        ]
      },
      "CreateCustomBotDto": {
        "type": "object",
        "properties": {
          "triggers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTriggerDto"
            }
          },
          "checkSchedule": {
            "type": "string"
          }
        },
        "required": [
          "triggers",
          "checkSchedule"
        ]
      },
      "CustomBot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "triggers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Trigger"
            }
          },
          "checkSchedule": {
            "type": "string"
          },
          "createdDate": {
            "format": "date-time",
            "type": "string"
          },
          "updateDate": {
            "format": "date-time",
            "type": "string"
          },
          "deletedDate": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "ownerId",
          "triggers",
          "checkSchedule",
          "createdDate",
          "updateDate",
          "deletedDate"
        ]
      },
      "UpdateTriggerDto": {
        "type": "object",
        "properties": {
          "updateSchedule": {
            "type": "string"
          },
          "triggerType": {
            "$ref": "#/components/schemas/TriggerTypeEnum"
          },
          "meta": {
            "type": "object"
          },
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "updateSchedule",
          "triggerType",
          "meta",
          "id",
          "uuid"
        ]
      },
      "UpdateCustomBotDto": {
        "type": "object",
        "properties": {
          "triggers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateTriggerDto"
            }
          },
          "checkSchedule": {
            "type": "string"
          }
        },
        "required": [
          "triggers",
          "checkSchedule"
        ]
      }
    }
  },
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "Application Support"
        ]
      }
    },
    "/authorise": {
      "get": {
        "operationId": "AppController_getHelloAuthorized",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "Application Support"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/organisation/{uuid}": {
      "get": {
        "operationId": "OrganisationController_findOne",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          }
        },
        "tags": [
          "organisations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "OrganisationController_update",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganisationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          }
        },
        "tags": [
          "organisations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/organisation": {
      "get": {
        "operationId": "OrganisationController_findAllForUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organisation"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "organisations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/organisation/uuid": {
      "delete": {
        "operationId": "OrganisationController_remove",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organisation"
                }
              }
            }
          }
        },
        "tags": [
          "organisations"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/person": {
      "get": {
        "operationId": "PersonController_findSelf",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "persons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/person/{uuid}": {
      "patch": {
        "operationId": "PersonController_update",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "persons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PersonController_remove",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Person"
                }
              }
            }
          }
        },
        "tags": [
          "persons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/email-client/verify": {
      "get": {
        "operationId": "EmailClientController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "email-client"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/custom-bot": {
      "post": {
        "operationId": "CustomBotController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomBotDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBot"
                }
              }
            }
          }
        },
        "tags": [
          "Custom Bot"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CustomBotController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomBot"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Custom Bot"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/custom-bot/trigger/{uuid}": {
      "post": {
        "operationId": "CustomBotController_trigger",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Custom Bot"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/custom-bot/{uuid}": {
      "get": {
        "operationId": "CustomBotController_findOne",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBot"
                }
              }
            }
          }
        },
        "tags": [
          "Custom Bot"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "CustomBotController_update",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomBotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomBot"
                }
              }
            }
          }
        },
        "tags": [
          "Custom Bot"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CustomBotController_remove",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Custom Bot"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  }
}