# shared-api-client/src/models/Trigger.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Coin bot Api
 * Describes the backend api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    TriggerResult,
    TriggerResultFromJSON,
    TriggerResultFromJSONTyped,
    TriggerResultToJSON,
    TriggerTypeEnum,
    TriggerTypeEnumFromJSON,
    TriggerTypeEnumFromJSONTyped,
    TriggerTypeEnumToJSON,
} from './';

/**
 * 
 * @export
 * @interface Trigger
 */
export interface Trigger {
    /**
     * 
     * @type {number}
     * @memberof Trigger
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Trigger
     */
    uuid: string;
    /**
     * 
     * @type {string}
     * @memberof Trigger
     */
    updateSchedule: string;
    /**
     * 
     * @type {TriggerTypeEnum}
     * @memberof Trigger
     */
    triggerType: TriggerTypeEnum;
    /**
     * 
     * @type {Array<TriggerResult>}
     * @memberof Trigger
     */
    triggerResults: Array<TriggerResult>;
    /**
     * 
     * @type {number}
     * @memberof Trigger
     */
    customBotId: number;
    /**
     * 
     * @type {object}
     * @memberof Trigger
     */
    meta: object;
    /**
     * 
     * @type {Date}
     * @memberof Trigger
     */
    createdDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof Trigger
     */
    updateDate: Date;
    /**
     * 
     * @type {Date}
     * @memberof Trigger
     */
    deletedDate: Date;
}

export function TriggerFromJSON(json: any): Trigger {
    return TriggerFromJSONTyped(json, false);
}

export function TriggerFromJSONTyped(json: any, ignoreDiscriminator: boolean): Trigger {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'uuid': json['uuid'],
        'updateSchedule': json['updateSchedule'],
        'triggerType': TriggerTypeEnumFromJSON(json['triggerType']),
        'triggerResults': ((json['triggerResults'] as Array<any>).map(TriggerResultFromJSON)),
        'customBotId': json['customBotId'],
        'meta': json['meta'],
        'createdDate': (new Date(json['createdDate'])),
        'updateDate': (new Date(json['updateDate'])),
        'deletedDate': (new Date(json['deletedDate'])),
    };
}

export function TriggerToJSON(value?: Trigger | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'uuid': value.uuid,
        'updateSchedule': value.updateSchedule,
        'triggerType': TriggerTypeEnumToJSON(value.triggerType),
        'triggerResults': ((value.triggerResults as Array<any>).map(TriggerResultToJSON)),
        'customBotId': value.customBotId,
        'meta': value.meta,
        'createdDate': (value.createdDate.toISOString()),
        'updateDate': (value.updateDate.toISOString()),
        'deletedDate': (value.deletedDate.toISOString()),
    };
}